+++ example/webpack.main.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  entry: './src/main.js',
+  module: {
+    rules: require('./webpack.rules'),
+  },
+}