+++ src/apis/Linking.js
@@ -1,6 +1,6 @@
 // @flow
 
-import { remote, shell } from 'electron'
+const remote = window.ReactNativeElectron
 
 const eventHandlers = new Map()
 
@@ -10,7 +10,7 @@ export const addEventListener = (type: string, handler: Function) => {
       handler({ type, url })
     }
     eventHandlers.set(handler, wrapHandler)
-    remote.app.on('open-url', wrapHandler)
+    remote.ipc.on('app-open-url', wrapHandler)
   }
 }
 
@@ -18,21 +18,19 @@ export const removeEventListener = (type: string, handler: Function) => {
   if (type === 'url' && typeof handler === 'function') {
     const wrapHandler = eventHandlers.get(handler)
     if (wrapHandler) {
-      remote.app.removeListener('open-url', wrapHandler)
+      remote.ipc.off('app-open-url', wrapHandler)
     }
     eventHandlers.delete(handler)
   }
 }
 
-export const openURL = (url: string, options: ?Object): Promise<void> => {
-  return shell.openExternal(url, options)
-    ? Promise.resolve()
-    : Promise.reject(new Error('Could not open URL'))
+export const openURL = async (url: string): Promise<void> => {
+  await remote.openURL(url)
 }
 
 // Apply same behavior as RNW
 export const canOpenURL = (): Promise<boolean> => Promise.resolve(true)
 
-export const getInitialURL = (): Promise<?string> => {
-  return Promise.resolve(remote.process.argv[1] || null)
+export const getInitialURL = async (): Promise<?string> => {
+  return (await remote.getInitialURL()) || null
 }